+++ R/3-shared-getMatrix.R
@@ -8,8 +8,15 @@
 #' @return A matrix.
 #'
 #' @export
-getMatrix <-
-  function(object) {
+getMatrix <- function(object) {
+
+  if(class(object) == "propr"){
     df <- object@matrix
-    return(df)
+  }else if(class(object) == "propd"){
+    df <- get_theta_matrix(object)
+  }else{
+    stop("Provided 'object' not recognized.")
+  }
+
+  return(df)
   }